Revert "[Homepage Banner] Convert to React component (#11818)"
This reverts commit 104c47f4f4d70df8e11c3404fdd5190a3d837877.
