--- src/applications/static-pages/homepage-banner/createHomepageBanner.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-export default function createHomepageBanner(store, widgetType) {
-  const root = document.querySelector(`[data-widget-type="${widgetType}"]`);
-  if (root) {
-    import(/* webpackChunkName: "HomepageBanner" */
-    './HomepageBanner').then(module => {
-      const HomepageBanner = module.default;
-      ReactDOM.render(<HomepageBanner />, root);
-    });
-  }
-}
