+++ src/applications/static-pages/widgetTypes.js
@@ -9,5 +9,4 @@ export default {
   DISABILITY_RATING_CALCULATOR: 'disability-rating-calculator',
   FIND_VA_FORMS: 'find-va-forms',
   HIGHER_LEVEL_REVIEW_APP_STATUS: 'higher-level-review-status',
-  HOMEPAGE_BANNER: 'homepage-banner',
 };
